fix(newapplicationmodal): rename startnewapplicationmodal to newapplicationmodal
