+++ source/components/organisms/NewApplicationModal/NewApplicationModal.styled.ts
