+++ storybook/storyLoader.js
@@ -42,8 +42,8 @@ function loadStories() {
   require("../source/components/organisms/BulletList/BulletList.stories");
   require("../source/components/organisms/CheckboxList/CheckboxList.stories");
   require("../source/components/organisms/FormList/FormList.stories");
+  require("../source/components/organisms/NewApplicationModal/NewApplicationModal.stories");
   require("../source/components/organisms/PinInputModal/PinInputModal.stories");
-  require("../source/components/organisms/StartNewApplicationModal/StartNewApplicationModal.stories");
   require("../source/components/organisms/Step/Step.stories");
   require("../source/components/organisms/Step/StepBanner/StepBanner.stories");
   require("../source/components/organisms/Step/StepDescription/StepDescription.stories");
@@ -95,8 +95,8 @@ const stories = [
   "../source/components/organisms/BulletList/BulletList.stories",
   "../source/components/organisms/CheckboxList/CheckboxList.stories",
   "../source/components/organisms/FormList/FormList.stories",
+  "../source/components/organisms/NewApplicationModal/NewApplicationModal.stories",
   "../source/components/organisms/PinInputModal/PinInputModal.stories",
-  "../source/components/organisms/StartNewApplicationModal/StartNewApplicationModal.stories",
   "../source/components/organisms/Step/Step.stories",
   "../source/components/organisms/Step/StepBanner/StepBanner.stories",
   "../source/components/organisms/Step/StepDescription/StepDescription.stories",
